[parser] folder structure for VariableStatement and ExpressionStatement
